Added cart page and update code
