+++ src/LandingPage/Page.css
@@ -1,0 +1,104 @@
+/* .model {
+  width: 100%;
+  height: 80vh;
+} */
+.models {
+  /* width: 100%; */
+  /* height: 65vh;
+  background-image: url("../Img/IMG_6126.jpg");
+  background-size: cover;
+  background-repeat: no-repeat;
+  animation-name: example;
+  animation-duration: 8s;
+  animation-iteration-count: infinite; */
+}
+
+/* @keyframes example {
+  0% {
+    background-image: url("../Img/IMG_6126.jpg");
+  }
+  25% {
+    background-image: url("../Img/IMG_6125.jpg");
+  }
+  50% {
+    background-image: url("../Img/IMG_6127.jpg");
+  }
+  100% {
+    background-image: url("../Img/IMG_6126.jpg");
+  }
+}
+.model-label h1 {
+  color: white;
+  text-align: center;
+  font-family: Montserrat;
+  font-size: 30px;
+  font-style: normal;
+  font-weight: 500;
+  letter-spacing: 2px;
+  text-transform: uppercase;
+  line-height: 50px;
+} */
+/* .model-label h4 {
+  color: white;
+  text-align: center;
+  font-family: Montserrat;
+  font-size: 15px;
+  font-style: normal;
+  font-weight: 500;
+  letter-spacing: 2px;
+  text-transform: uppercase;
+}
+.model-label {
+  padding-top: 250px;
+} */
+/* @media (min-width: 768px) {
+  .models {
+    background-image: url("../Img/modelTab1.svg");
+    background-size: contain;
+    background-repeat: no-repeat;
+    animation-name: example;
+    animation-duration: 8s;
+    animation-iteration-count: infinite;
+  }
+  
+  @keyframes example {
+    0% {
+      background-image: url("../Img/modelTab1.svg");
+    }
+    25% {
+      background-image: url("../Img/modelTab2.svg");
+    }
+    50% {
+      background-image: url("../Img/modelTab3.svg");
+    }
+    100% {
+      background-image: url("../Img/modelTab4.svg");
+    }
+  }
+} */
+@media (min-width: 1024px) {
+  /* .models {
+    background-image: url("../Img/model1.svg");
+    background-size: cover;
+    background-repeat: no-repeat;
+    animation-name: example;
+    width: 100%;
+    animation-duration: 8s ease-out;
+    animation-iteration-count: infinite;
+  }
+  
+  @keyframes example {
+    0% {
+      background-image: url("../Img/model1.svg");
+    }
+    25% {
+      background-image: url("../Img/model5.svg");
+    }
+    0% {
+      background-image: url("../Img/model2.svg");
+    }
+    50% {
+      background-image: url("../Img/model1.svg");
+    }
+  } */
+}
