+++ src/LandingPage/Page.jsx
@@ -1,0 +1,13 @@
+import React from "react";
+import "./Page1.css";
+
+
+const Page1 = () => {
+  return (
+    <div>
+        <h1>PABLO DIMENSION</h1>
+    </div>
+  );
+};
+
+export default Page1;
