+++ src/Navbar/Navbar.css
@@ -2,12 +2,16 @@
   padding: 0;
   margin: 0;
 }
-nav {
+/* nav {
   position: fixed;
   background-color: #333;
-}
+} */
 .navBar {
   position: relative;
+  position: -webkit-sticky;
+  position: sticky;
+  top: 0px;
+  z-index: 1;
 }
 .pabloHead {
   background-color: black;
@@ -24,9 +28,9 @@ nav {
   padding: 15px 0;
 }
 .Nav {
-  /* position: -webkit-sticky; */
+  position: -webkit-sticky;
   position: sticky;
-  top: 0; 
+  top: 0px; 
   z-index: 1;
   background-color: white;
   padding: 7px 0;
@@ -240,7 +244,7 @@ nav {
     justify-content: center;
     gap: 40px;
   }
-  .Nav img:nth-child(4) {
+  .Nav .Link:nth-child(4) {
     order: -1;
   }
   .hamburger {
