+++ src/Shop-details/Latestdetails.jsx
@@ -31,66 +31,75 @@ const Latestdetails = () => {
   };
   return (
     <div className="latest-details">
-      <div className="latest-div1">
-        <Carousel responsive={responsive} className="carousel-latest">
-          <div className="details">
-            <img src={shoe1} alt="" className="latest-details-img1" />
-          </div>
-          <div>
-            <img src={shoe2} alt="" className="latest-details-img1" />
-          </div>
-          <div>
-            <img src={shoe3} alt="" className="latest-details-img1" />
-          </div>
-          <div>
-            <img src={shoe4} alt="" className="latest-details-img1" />
-          </div>
-          <div>
-            <img src={shoe5} alt="" className="latest-details-img1" />
-          </div>
-        </Carousel>
-      </div>
-      <div className="latest-div2">
-        <h1 className="latest-name">Pablo Dimension female Joggers</h1>
-        <p className="latest-price">$8,999</p>
-        <div className="size">
-          <h2 className="size-txt">Size</h2>
-          <div className="size-box">
-            <p>M</p>
-            <p>L</p>
-            <p>XL</p>
-          </div>
+      <div className="latest-display">
+        <div className="latest-div1">
+          <Carousel
+            responsive={responsive}
+            className="carousel-latest"
+            swipeable={true}
+            draggable={true}
+            showDots={false}
+            renderDotsOutside={false}
+          >
+            <div className="details">
+              <img src={shoe1} alt="" className="latest-details-img1" />
+            </div>
+            <div>
+              <img src={shoe2} alt="" className="latest-details-img1" />
+            </div>
+            <div>
+              <img src={shoe3} alt="" className="latest-details-img1" />
+            </div>
+            <div>
+              <img src={shoe4} alt="" className="latest-details-img1" />
+            </div>
+            <div>
+              <img src={shoe5} alt="" className="latest-details-img1" />
+            </div>
+          </Carousel>
         </div>
-        <div className="color">
-          <h2 className="color-txt">Color</h2>
-          <div className="color-box">
-            <p>white</p>
-            <p>black</p>
-            <p>orange</p>
-            <p>green</p>
-          </div>
-        </div>
-        <div>
-          <h2 className="description-head">PRODUCT DETAILS</h2>
-          <div>
-            <div className="description">
-              <h3 className="description-txt">Description</h3>
-              <h2 className="description-name">ALEXANDER MCQUEEN</h2>
-              <p className="description-p">
-                Brushstroke Harness abstract-print shirt For AW23, Alexander
-                McQueen puts a futuristic spin on sartorial silhouettes, as
-                exemplified by this shirt. Cut from stretch cotton poplin, a
-                contrasting harness detailed with the ‘Brushstroke’ print
-                interrupts the conventional collared style and gives it a unique
-                edge.
-              </p>
+        <div className="latest-div2">
+          <h1 className="latest-name">Pablo Dimension female Joggers</h1>
+          <p className="latest-price">$8,999</p>
+          <div className="size">
+            <h2 className="size-txt">Size</h2>
+            <div className="size-box">
+              <p>M</p>
+              <p>L</p>
+              <p>XL</p>
             </div>
           </div>
+          <div className="color">
+            <h2 className="color-txt">Color</h2>
+            <div className="color-box">
+              <p>white</p>
+              <p>black</p>
+              <p>orange</p>
+              <p>green</p>
+            </div>
+          </div>
+          <div>
+            <h2 className="description-head">PRODUCT DETAILS</h2>
+            <div>
+              <div className="description">
+                <h3 className="description-txt">Description</h3>
+                <h2 className="description-name">ALEXANDER MCQUEEN</h2>
+                <p className="description-p">
+                  Brushstroke Harness abstract-print shirt For AW23, Alexander
+                  McQueen puts a futuristic spin on sartorial silhouettes, as
+                  exemplified by this shirt. Cut from stretch cotton poplin, a
+                  contrasting harness detailed with the ‘Brushstroke’ print
+                  interrupts the conventional collared style and gives it a
+                  unique edge.
+                </p>
+              </div>
+            </div>
+          </div>
+          <div className="button-cart">
+            <button className="call">CA</button>
+            <button className="add-cart">ADD TO CART</button>
+          </div>
         </div>
-      </div>
-      <div className="button-cart">
-        <button className="call">CA</button>
-        <button className="add-cart">ADD TO CART</button>
       </div>
     </div>
   );
