+++ src/Shop/Latest.css
@@ -79,9 +79,15 @@
 .empty-div {
   display: none;
 }
+.Link {
+  text-decoration: none;
+}
 @media (min-width: 540px) {
   .grid-img {
     height: 330px;
+  }
+  .display-box{
+    padding: 40px 20px;
   }
 }
 @media (min-width: 767px) {
@@ -107,9 +113,15 @@
   .grid-box {
     padding: 0px 30px;
   }
+  .display-box{
+    padding: 40px 20px;
+  }
 }
-@media (min-width: 767px) {
+@media (min-width: 1024px) {
   .display-box {
     grid-template-columns: 1fr 1fr 1fr 1fr;
+  }
+  .display-box{
+    padding: 50px 30px;
   }
 }
